+++ release_notes.md
@@ -13,3 +13,4 @@
 - Fixed an issue leading to a race when invocation responses returned prior to HTTP requests being sent in proxied scenarios.
 - Language worker channels will not be started during placeholder mode if we are in-process (#10161)
 - Ordered invocations are now the default (#10201)
+- Fixed incorrect function count in the log message.(#10220)